fix: pre-rendering props is now fixed
Props were not being passed properly because initial setup had 'res' response being parsed as undefined. This had to do with returning an object right off the bat
Set the object up as a variable and then returned the variable. This fixed the 'undefined' issue. Had to change some of server.js code ot acoomodate, but works excellently now.